refactor: consolidate utils and lib into shared/ for clearer directory roles
src/utils/ と src/lib/ を src/shared/ 配下に統合し、
src/ 直下のディレクトリの役割を明確化。

- src/utils/algorithm.ts → src/shared/utils/algorithm.ts
- src/lib/i18n.ts → src/shared/lib/i18n.ts
- import パスを更新（影響2ファイル）

src/ 直下の構成:
  features/   feature-scoped UI
  shared/     cross-feature（components / hooks / utils / lib）
  services/   service layer
  entrypoints/ エントリポイント
