+++ src/shared/hooks/useTranslation.ts
@@ -1,4 +1,4 @@
-import { i18n } from "@/lib/i18n";
+import { i18n } from "@/shared/lib/i18n";
 
 export const useTranslation = () => {
   return {
