+++ src/shared/utils/algorithm.ts
